项目列表接口新增分页查询，去除原有全量查询接口

+++ src/main/java/com/dagu/lightchaser/entity/PageParamEntity.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class PageParamEntity {
-    private int size;
-    private int current;
+    private Long size;
+    private Long current;
     private String searchValue;
 }
